refactor(backend): Refined Vehicle update and delete logic to manage ownership and associated rides

+++ wayzi-spring-react/wayzi-backend/src/main/java/mk/ukim/finki/wayzi/service/domain/impl/VehicleServiceImpl.java
@@ -4,12 +4,15 @@ import mk.ukim.finki.wayzi.dto.CreateVehicleDto;
 import mk.ukim.finki.wayzi.dto.DisplayVehicleDto;
 import mk.ukim.finki.wayzi.exception.AccessDeniedException;
 import mk.ukim.finki.wayzi.exception.VehicleNotFoundException;
+import mk.ukim.finki.wayzi.model.domain.ride.Ride;
 import mk.ukim.finki.wayzi.model.domain.user.StandardUser;
 import mk.ukim.finki.wayzi.model.domain.user.User;
 import mk.ukim.finki.wayzi.model.domain.vehicle.Vehicle;
 import mk.ukim.finki.wayzi.repository.VehicleRepository;
 import mk.ukim.finki.wayzi.service.domain.AuthService;
+import mk.ukim.finki.wayzi.service.domain.RideService;
 import mk.ukim.finki.wayzi.service.domain.VehicleService;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,10 +22,12 @@ public class VehicleServiceImpl implements VehicleService {
 
     private final VehicleRepository vehicleRepository;
     private final AuthService authService;
+    private final RideService rideService;
 
-    public VehicleServiceImpl(VehicleRepository vehicleRepository, AuthService authService) {
+    public VehicleServiceImpl(VehicleRepository vehicleRepository, AuthService authService, @Lazy RideService rideService) {
         this.vehicleRepository = vehicleRepository;
         this.authService = authService;
+        this.rideService = rideService;
     }
 
     @Override
@@ -62,19 +67,32 @@ public class VehicleServiceImpl implements VehicleService {
     @Override
     public Vehicle update(Long id, CreateVehicleDto createVehicleDto) {
         Vehicle vehicle = findByIdAndCheckOwnership(id);
+        List<Ride> rides = rideService.findAllForAuthenticatedUserByVehicleId(id);
 
-        vehicle.setType(createVehicleDto.type());
-        vehicle.setCapacity(createVehicleDto.capacity());
-        vehicle.setModel(createVehicleDto.model());
-        vehicle.setBrand(createVehicleDto.brand());
-        vehicle.setColor(createVehicleDto.color());
-
-        return vehicleRepository.save(vehicle);
+        if(rides.isEmpty()) {
+            vehicle.setType(createVehicleDto.type());
+            vehicle.setCapacity(createVehicleDto.capacity());
+            vehicle.setModel(createVehicleDto.model());
+            vehicle.setBrand(createVehicleDto.brand());
+            vehicle.setColor(createVehicleDto.color());
+            return vehicleRepository.save(vehicle);
+        } else {
+            vehicle.setOwner(null);
+            vehicleRepository.save(vehicle);
+            return save(createVehicleDto);
+        }
     }
 
     @Override
     public void delete(Long id) {
         Vehicle vehicle = findByIdAndCheckOwnership(id);
-        vehicleRepository.delete(vehicle);
+
+        List<Ride> rides = rideService.findAllForAuthenticatedUserByVehicleId(id);
+        if(rides.isEmpty()) {
+            vehicleRepository.delete(vehicle);
+        } else {
+            vehicle.setOwner(null);
+            vehicleRepository.save(vehicle);
+        }
     }
 }
